+++ fis_police_server/src/main/java/fis/police/fis_police_server/repository/AgentRepository.java
@@ -14,5 +14,5 @@ public interface AgentRepository {
 
     void save(Agent agent);
 
-    List<Agent> findNearAgent(Float latitude, Float longitude, Long range);
+    List<Agent> findNearAgent(Double latitude, Double longitude, Long range);
 }
